% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_pct}
\alias{format_pct}
\title{Format a proportion as percentage string}
\usage{
format_pct(x)
}
\arguments{
\item{x}{numeric between 0 and 1}
}
\value{
character string like "45.2\%"
}
\description{
Format a proportion as percentage string
}
\keyword{internal}
