% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_qc_distributions}
\alias{plot_qc_distributions}
\title{Plot QC metric distributions}
\usage{
plot_qc_distributions(qc_metrics, pass)
}
\arguments{
\item{qc_metrics}{data.frame from compute_qc_metrics()}

\item{pass}{Logical vector (TRUE = passes)}
}
\value{
Named list of ggplot2 objects
}
\description{
Returns a list of individual ggplot2 histograms, one per metric.
}
\keyword{internal}
