% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umap_qc}
\alias{plot_umap_qc}
\title{Plot UMAP with pass/fail coloring}
\usage{
plot_umap_qc(obsm, pass, reduction = "X_umap")
}
\arguments{
\item{obsm}{Named list of reduced dimensions (from schard list)}

\item{pass}{Logical vector (TRUE = passes QC)}

\item{reduction}{Character, which reduction to plot (default "X_umap")}
}
\value{
ggplot2 object
}
\description{
Plot UMAP with pass/fail coloring
}
\keyword{internal}
